requirements.txt
google-genai==1.65.0
python-dotenv==1.2.2
fastapi==0.135.1
uvicorn==0.41.0